+++ messages/Makefile
@@ -1,2 +1,3 @@
-protobuf: mockMsg.proto
+protobuf: mockMsg.proto ctrlMsg.proto
 	protoc -I=./ --python_out=./ ./mockMsg.proto
+	protoc -I=./ --python_out=./ ./ctrlMsg.proto
